% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/main.R
\name{drm.comparable.model}
\alias{drm.comparable.model}
\title{Build model for data set with two or more samples}
\usage{
drm.comparable.model(df)
}
\arguments{
\item{df}{- input data.frame in long format with two or more samples}
}
\value{
The object of class \code{drc}
}
\description{
Build model for data set with two or more samples
}

