% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ModelMaker.R
\name{drm.single.model}
\alias{drm.single.model}
\title{Build model for data set with single sample}
\usage{
drm.single.model(df)
}
\arguments{
\item{df}{- input data.frame in long format}
}
\value{
The object of class \code{drc}
}
\description{
Build model for data set with single sample
}

